#327 [feat] ImgUrl을 통해 Bitmap 가져와 사용하는 UtilFunction 정의 후 리팩토링
